Move blob from compiler/ to util/
There's nothing whatsoever compiler-specific about it other than that's
currently where it's used.
