+++ src/util/blob.c
