+++ src/util/blob.h
