+++ src/util/blob_test.c
